test: run ceph test separately to realistic test and add some documentation
